backend/requirements.txt
fastapi
uvicorn
sqlalchemy
bcrypt>=4.0.0
python-jose[cryptography]
python-multipart
pydantic[email]
fastapi-mail
python-dotenv
langchain
langchain-openai
langchain-community
chromadb
pypdf
docx2txt
tiktoken


